VFEP-83 rename nightTimePhone to mobilePhone and dayTimePhone to homePhone. Switch the order of mobile phone and home phone.
